Fix proxy renaming to middleware
