+++ middleware.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, afterEach, describe, expect, it, vi } from 'vitest';
 import { NextRequest, NextResponse } from 'next/server';
-import { proxy, config } from './proxy';
+import { middleware, config } from './middleware';
 import { rateLimit } from '@/lib/rate-limit';
 
 vi.mock('@/lib/rate-limit', () => ({
@@ -19,7 +19,7 @@ function mockBothLimiters(
   });
 }
 
-describe('proxy', () => {
+describe('middleware', () => {
   let originalEnv: string | undefined;
 
   beforeEach(() => {
@@ -43,7 +43,7 @@ describe('proxy', () => {
     const nextSpy = vi.spyOn(NextResponse, 'next');
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    await proxy(request);
+    await middleware(request);
 
     expect(nextSpy).toHaveBeenCalled();
   });
@@ -57,7 +57,7 @@ describe('proxy', () => {
     });
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    const response = await proxy(request);
+    const response = await middleware(request);
 
     expect(response.status).toBe(429);
   });
@@ -71,7 +71,7 @@ describe('proxy', () => {
     });
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    const response = await proxy(request);
+    const response = await middleware(request);
 
     await expect(response.json()).resolves.toEqual({
       error: 'Too many requests',
@@ -87,7 +87,7 @@ describe('proxy', () => {
     });
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    await proxy(request);
+    await middleware(request);
 
     expect(rateLimit).toHaveBeenCalledWith(expect.any(String), 60, 60000);
   });
@@ -101,7 +101,7 @@ describe('proxy', () => {
     });
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    const response = await proxy(request);
+    const response = await middleware(request);
 
     expect(response.headers.get('X-RateLimit-Limit')).toBe('60');
     expect(response.headers.get('X-RateLimit-Remaining')).toBe('59');
@@ -117,7 +117,7 @@ describe('proxy', () => {
     });
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    const response = await proxy(request);
+    const response = await middleware(request);
 
     expect(response.headers.get('X-RateLimit-Limit')).toBe('60');
     expect(response.headers.get('X-RateLimit-Remaining')).toBe('58');
@@ -133,7 +133,7 @@ describe('proxy', () => {
     });
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    const response = await proxy(request);
+    const response = await middleware(request);
 
     expect(response.headers.get('Content-Type')).toBe('application/json');
     expect(response.headers.get('X-RateLimit-Limit')).toBe('60');
@@ -155,7 +155,7 @@ describe('proxy', () => {
       },
     });
 
-    await proxy(request);
+    await middleware(request);
 
     expect(rateLimit).toHaveBeenCalledWith('1.2.3.4', 60, 60000);
   });
@@ -176,7 +176,7 @@ describe('proxy', () => {
       },
     });
 
-    await proxy(request);
+    await middleware(request);
 
     expect(rateLimit).toHaveBeenCalledWith('5.6.7.8', 60, 60000);
   });
@@ -195,7 +195,7 @@ describe('proxy', () => {
       },
     });
 
-    await proxy(request);
+    await middleware(request);
 
     expect(rateLimit).toHaveBeenCalledWith('9.9.9.9', 60, 60000);
   });
@@ -210,7 +210,7 @@ describe('proxy', () => {
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
 
-    await proxy(request);
+    await middleware(request);
 
     expect(rateLimit).toHaveBeenCalledWith('127.0.0.1', 60, 60000);
   });
@@ -230,7 +230,7 @@ describe('proxy', () => {
       },
     });
 
-    await proxy(request);
+    await middleware(request);
 
     expect(rateLimit).toHaveBeenCalledWith('1.2.3.4', 60, 60000);
   });
@@ -249,7 +249,7 @@ describe('proxy', () => {
       },
     });
 
-    await proxy(request);
+    await middleware(request);
 
     expect(rateLimit).toHaveBeenCalledWith('1.2.3.4', 60, 60000);
   });
@@ -268,7 +268,7 @@ describe('proxy', () => {
       mockBothLimiters({ success: true, limit: 5, remaining: 4, reset: 123456789 });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      await proxy(request);
+      await middleware(request);
 
       expect(rateLimit).toHaveBeenNthCalledWith(1, 'refresh:127.0.0.1', 5, 60000);
       expect(rateLimit).toHaveBeenNthCalledWith(2, '127.0.0.1', 60, 60000);
@@ -278,7 +278,7 @@ describe('proxy', () => {
       mockBothLimiters({ success: false, limit: 5, remaining: 0, reset: 123456789 });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      const response = await proxy(request);
+      const response = await middleware(request);
 
       expect(response.status).toBe(429);
       const body = await response.json();
@@ -294,7 +294,7 @@ describe('proxy', () => {
       });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      const response = await proxy(request);
+      const response = await middleware(request);
 
       expect(response.headers.get('X-RateLimit-Limit')).toBe('5');
       expect(response.status).toBe(429);
@@ -309,7 +309,7 @@ describe('proxy', () => {
       });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      await proxy(request);
+      await middleware(request);
 
       expect(rateLimit).not.toHaveBeenCalledWith('127.0.0.1', 60, 60000);
     });
@@ -323,7 +323,7 @@ describe('proxy', () => {
       });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      const response = await proxy(request);
+      const response = await middleware(request);
 
       expect(response.headers.get('X-RateLimit-Limit')).toBe('5');
       expect(response.headers.get('X-RateLimit-Remaining')).toBe('0');
@@ -338,7 +338,7 @@ describe('proxy', () => {
       });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-      await proxy(request);
+      await middleware(request);
 
       expect(rateLimit).toHaveBeenCalledTimes(1);
       expect(rateLimit).toHaveBeenCalledWith('127.0.0.1', 60, 60000);
@@ -355,7 +355,7 @@ describe('proxy', () => {
       const request = new NextRequest(
         'http://localhost:3000/api/streak?user=octocat&refresh=false'
       );
-      await proxy(request);
+      await middleware(request);
 
       expect(rateLimit).toHaveBeenCalledTimes(1);
       expect(rateLimit).not.toHaveBeenCalledWith(expect.stringContaining('refresh:'), 5, 60000);
@@ -365,7 +365,7 @@ describe('proxy', () => {
       mockBothLimiters({ success: true, limit: 5, remaining: 3, reset: 123456789 });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      const response = await proxy(request);
+      const response = await middleware(request);
 
       expect(rateLimit).toHaveBeenCalledTimes(2);
       expect(response.status).toBe(200);
@@ -377,7 +377,7 @@ describe('proxy', () => {
         .mockResolvedValueOnce({ success: false, limit: 60, remaining: 0, reset: 123456789 });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      const response = await proxy(request);
+      const response = await middleware(request);
 
       expect(response.status).toBe(429);
       const body = await response.json();
@@ -388,14 +388,14 @@ describe('proxy', () => {
 
 describe('middleware.ts wiring', () => {
   it('middleware.ts exports a function named middleware', async () => {
-    const mod = await import('./proxy');
+    const mod = await import('./middleware');
 
     // Next.js looks for a named export called `middleware`
-    expect(typeof mod.proxy).toBe('function');
+    expect(typeof mod.middleware).toBe('function');
   });
 
   it('middleware.ts exports config with a non-empty matcher array', async () => {
-    const mod = await import('./proxy');
+    const mod = await import('./middleware');
 
     expect(mod.config).toBeDefined();
     expect(Array.isArray(mod.config.matcher)).toBe(true);
@@ -403,7 +403,7 @@ describe('middleware.ts wiring', () => {
   });
 
   it('middleware covers all expected API routes', async () => {
-    const { config: mwConfig } = await import('./proxy');
+    const { config: mwConfig } = await import('./middleware');
     const expected = [
       '/api/streak/:path*',
       '/api/github/:path*',
